Improving integer types in RPC code.
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_softfloat@1657 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/slirp/rpc/rpc.h
@@ -128,7 +128,7 @@ void rpc_uninit(void);
 
 void rpc_log(struct rpc_t* rpc, const char *format, ...);
 
-int rpc_read_file(const char* vfs_path, size_t offset, uint8_t* data, size_t len);
+int rpc_read_file(const char* vfs_path, uint32_t offset, uint8_t* data, uint32_t len);
 
 int rpc_match_arp(uint8_t byte);
 int rpc_match_icmp(uint32_t addr);
